oidc_child: fallback to ID and access tokens when looking up configured user identity
Some IdPs do not provide all attributes as requested by the claims in
userinfo response. They, however, provide them in either ID or access
tokens. Fall back to one of those if possible, in order to find the user
identity based on the attribute specified by the caller.

One example of such behavior is Entra ID:
https://learn.microsoft.com/en-us/entra/identity-platform/userinfo#userinfo-response

  > You can't add to or customize the information returned by the UserInfo endpoint.

Make sure we put the right token description when producing debug
information as well, to help seeing where the user identifier is coming
from.

Signed-off-by: Alexander Bokovoy <abokovoy@redhat.com>